Refactor some files into new Lists folder to reduce clutter
