+++ randomizer/Lists/WrinklyHints.py
