+++ randomizer/Spoiler.py
@@ -2,8 +2,8 @@
 
 import json
 
-from randomizer.Location import LocationList
-from randomizer.Item import ItemList
+from randomizer.Lists.Location import LocationList
+from randomizer.Lists.Item import ItemList
 from randomizer.ShuffleExits import ShufflableExits
